fix(app-first-load)/fixed app load using code splitting and lazy load
