+++ packages/bruno-app/src/components/ApiSpecPanel/Renderers/Swagger/index.js
@@ -1,5 +1,6 @@
 import { memo } from 'react';
 import SwaggerUI from 'swagger-ui-react';
+import 'swagger-ui-react/swagger-ui.css';
 import StyledWrapper from './StyledWrapper';
 import { serializeBody } from './serializeBody';
