+++ packages/bruno-app/src/components/RequestTabs/index.js
@@ -1,11 +1,12 @@
-import React, { useState, useRef, useEffect, useCallback, useMemo } from 'react';
+import React, { useState, useRef, useEffect, useCallback, useMemo, lazy, Suspense } from 'react';
 import find from 'lodash/find';
 import filter from 'lodash/filter';
 import classnames from 'classnames';
 import { IconChevronRight, IconChevronLeft } from '@tabler/icons';
 import { useSelector, useDispatch } from 'react-redux';
 import { focusTab, reorderTabs } from 'providers/ReduxStore/slices/tabs';
-import NewRequest from 'components/Sidebar/NewRequest';
+
+const NewRequest = lazy(() => import('components/Sidebar/NewRequest'));
 import CollectionHeader from './CollectionHeader';
 import RequestTab from './RequestTab';
 import StyledWrapper from './StyledWrapper';
@@ -111,7 +112,9 @@ const RequestTabs = () => {
   return (
     <StyledWrapper>
       {newRequestModalOpen && (
-        <NewRequest collectionUid={activeCollection?.uid} onClose={() => setNewRequestModalOpen(false)} />
+        <Suspense fallback={null}>
+          <NewRequest collectionUid={activeCollection?.uid} onClose={() => setNewRequestModalOpen(false)} />
+        </Suspense>
       )}
       {collectionRequestTabs && collectionRequestTabs.length ? (
         <>
