+++ packages/bruno-app/src/components/ResponsePane/ResponseBookmark/index.js
@@ -4,7 +4,8 @@ import { IconBookmark } from '@tabler/icons';
 import { addResponseExample } from 'providers/ReduxStore/slices/collections';
 import { saveRequest } from 'providers/ReduxStore/slices/collections/actions';
 import { insertTaskIntoQueue } from 'providers/ReduxStore/slices/app';
-import { uuid, formatResponse } from 'utils/common';
+import { uuid } from 'utils/common';
+import { formatResponse } from 'utils/common/format-response';
 import toast from 'react-hot-toast';
 import CreateExampleModal from 'components/ResponseExample/CreateExampleModal';
 import { getBodyType } from 'utils/responseBodyProcessor';
