+++ packages/bruno-app/src/pages/Main.js
@@ -10,19 +10,11 @@ import ThemeProvider from 'providers/Theme/index';
 import ErrorBoundary from './ErrorBoundary';
 
 import '../styles/globals.css';
-import 'codemirror/lib/codemirror.css';
-import 'graphiql/graphiql.min.css';
 import 'react-tooltip/dist/react-tooltip.css';
-import '@usebruno/graphql-docs/dist/esm/index.css';
-import '@fontsource/inter/100.css';
-import '@fontsource/inter/200.css';
-import '@fontsource/inter/300.css';
 import '@fontsource/inter/400.css';
 import '@fontsource/inter/500.css';
 import '@fontsource/inter/600.css';
 import '@fontsource/inter/700.css';
-import '@fontsource/inter/800.css';
-import '@fontsource/inter/900.css';
 import { setupPolyfills } from 'utils/common/setupPolyfills';
 setupPolyfills();
